-postgres helper functions
